feat: integrate Tailwind CSS and Angular Material; enhance authentication layout and styles
