+++ carebridge-admin/.postcssrc.json
@@ -1,0 +1,5 @@
+{
+    "plugins": {
+        "@tailwindcss/postcss": {}
+    }
+}
