+++ carebridge-admin/angular.json
@@ -27,6 +27,7 @@
               }
             ],
             "styles": [
+              "src/theme/custom-theme.scss",
               "src/styles.css"
             ],
             "scripts": [],
