+++ carebridge-admin/src/app/auth/auth-route.module.ts
@@ -1,0 +1,24 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+import { RegisterComponent } from '../register/register.component';
+import { AuthComponent } from './layouts/auth.component';
+import { LoginComponent } from './pages/login/login.component';
+
+const routes: Routes = [
+    {
+        path: '',
+        component: AuthComponent,
+        children: [
+            { path: '', redirectTo: 'login', pathMatch: 'full' },
+            { path: 'login', component: LoginComponent },
+            { path: 'register', component: RegisterComponent },
+            { path: '**', redirectTo: 'login' }
+        ]
+    }
+];
+
+@NgModule({
+    imports: [RouterModule.forChild(routes)],
+    exports: [RouterModule]
+})
+export class AuthRoutingModule { }
