+++ carebridge-admin/src/app/auth/layouts/auth.component.ts
@@ -1,8 +1,9 @@
 import { Component } from '@angular/core';
+import { RouterModule } from '@angular/router';
 
 @Component({
+  standalone: false,
   selector: 'app-auth',
-  imports: [],
   templateUrl: './auth.component.html',
   styleUrl: './auth.component.css'
 })
