+++ carebridge-admin/src/app/auth/pages/login/login.component.html
@@ -1,1 +1,2 @@
 <p>login works!</p>
+<mat-slide-toggle>Toggle me!</mat-slide-toggle>
