+++ carebridge-admin/src/app/auth/pages/login/login.component.ts
@@ -1,13 +1,16 @@
 import { Component } from '@angular/core';
 
 @Component({
+  standalone: false,
   selector: 'app-login',
-  imports: [],
   templateUrl: './login.component.html',
-  styleUrl: './login.component.css'
+  styleUrls: ['./login.component.css']
 })
-
-
 export class LoginComponent {
+  username = '';
+  password = '';
 
+  login() {
+    console.log('Logging in with', this.username, this.password);
+  }
 }
