+++ carebridge-admin/src/theme/_theme-colors.scss
@@ -1,0 +1,137 @@
+// This file was generated by running 'ng generate @angular/material:theme-color'.
+// Proceed with caution if making changes to this file.
+
+@use 'sass:map';
+@use '@angular/material' as mat;
+
+// Note: Color palettes are generated from primary: #34A853, secondary: #59C877, tertiary: #F0F8EE, neutral: #f0f8ee80
+$_palettes: (
+  primary: (
+    0: #000000,
+    10: #002108,
+    20: #003914,
+    25: #004619,
+    30: #005320,
+    35: #006026,
+    40: #006e2c,
+    50: #008a3a,
+    60: #30a550,
+    70: #50c168,
+    80: #6ddd81,
+    90: #89fa9b,
+    95: #c6ffc7,
+    98: #ebffe7,
+    99: #f6fff1,
+    100: #ffffff,
+  ),
+  secondary: (
+    0: #000000,
+    10: #00210b,
+    20: #003917,
+    25: #00461d,
+    30: #005224,
+    35: #00602b,
+    40: #006d32,
+    50: #008a41,
+    60: #32a558,
+    70: #51c070,
+    80: #6edd89,
+    90: #8afaa3,
+    95: #c4ffcb,
+    98: #eaffe9,
+    99: #f5fff2,
+    100: #ffffff,
+  ),
+  tertiary: (
+    0: #000000,
+    10: #161d17,
+    20: #2b322c,
+    25: #363d37,
+    30: #414942,
+    35: #4d544d,
+    40: #586059,
+    50: #717971,
+    60: #8b938a,
+    70: #a5ada4,
+    80: #c1c9bf,
+    90: #dde5db,
+    95: #ebf3e9,
+    98: #f4fcf2,
+    99: #f7fef4,
+    100: #ffffff,
+  ),
+  neutral: (
+    0: #000000,
+    10: #1f1c00,
+    20: #353100,
+    25: #413d00,
+    30: #4d4800,
+    35: #5a5400,
+    40: #676000,
+    50: #817916,
+    60: #9b9330,
+    70: #b7ae48,
+    80: #d3ca60,
+    90: #f0e679,
+    95: #fef485,
+    98: #fffae2,
+    99: #fffbff,
+    100: #ffffff,
+    4: #100f00,
+    6: #161400,
+    12: #232000,
+    17: #2e2b00,
+    22: #3a3600,
+    24: #3f3a00,
+    87: #e7dd71,
+    92: #f5ec7e,
+    94: #fbf183,
+    96: #fff6a2,
+  ),
+  neutral-variant: (
+    0: #000000,
+    10: #131e14,
+    20: #283328,
+    25: #333e33,
+    30: #3e4a3e,
+    35: #4a5549,
+    40: #566255,
+    50: #6e7a6d,
+    60: #889486,
+    70: #a2af9f,
+    80: #bdcaba,
+    90: #d9e6d6,
+    95: #e8f4e4,
+    98: #f0fdec,
+    99: #f6fff1,
+    100: #ffffff,
+  ),
+  error: (
+    0: #000000,
+    10: #410002,
+    20: #690005,
+    25: #7e0007,
+    30: #93000a,
+    35: #a80710,
+    40: #ba1a1a,
+    50: #de3730,
+    60: #ff5449,
+    70: #ff897d,
+    80: #ffb4ab,
+    90: #ffdad6,
+    95: #ffedea,
+    98: #fff8f7,
+    99: #fffbff,
+    100: #ffffff,
+  ),
+);
+
+$_rest: (
+  secondary: map.get($_palettes, secondary),
+  neutral: map.get($_palettes, neutral),
+  neutral-variant: map.get($_palettes,  neutral-variant),
+  error: map.get($_palettes, error),
+);
+
+$primary-palette: map.merge(map.get($_palettes, primary), $_rest);
+$tertiary-palette: map.merge(map.get($_palettes, tertiary), $_rest);
