+++ carebridge-admin/src/theme/custom-theme.scss
@@ -1,0 +1,19 @@
+// Custom Theming for Angular Material
+// For more information: https://material.angular.io/guide/theming
+@use '@angular/material' as mat;
+@use './_theme-colors' as theme-colors;
+
+html {
+  @include mat.theme((
+    color: (
+      primary: theme-colors.$primary-palette,
+      // secondary: theme-colors.$secondary-palette,
+      tertiary: theme-colors.$tertiary-palette,
+      // error: theme-colors.$error-palette,
+      // neutral: theme-colors.$neutral-palette,
+      // 'neutral-variant': theme-colors.$neutral-variant-palette,
+    ),
+    typography: Roboto,
+    density: -2,
+  ));
+}
